refactor(store-core): extract error handlers into handlers/error.ts (audit P2-3)
Move the error family out of the handlers/index.ts god-module into
handlers/error.ts, behind the barrel's export * re-export so the public surface
is unchanged:

- Client error envelope: handleError + DEFAULT_ERROR_MESSAGE +
  pickFiniteTokenCount (the #5039 partial-cost extraction)
- Server-lifecycle errors: handleServerError + SERVER_ERROR_CATEGORIES,
  handleServerShutdown, handleServerStatusLegacy + their payload interfaces

Drop the now-unused ServerError (../types), ErrorPartialCost (../cost-format),
and stripAnsi (../utils) imports from index.ts — all only used by the moved
handlers. ChatMessage and nextMessageId stay (still used by permission /
user-question handlers).

Pure move — both ranges byte-identical to origin/main. index.ts 1807 -> 1562 lines.
